lambda/api/requirements.txt
psycopg2-binary>=2.9
boto3>=1.34
google-genai>=1.0
